requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
faiss-cpu==1.12.0
opencv-python==4.8.1.78
numpy<2.0,>=1.24.0
pydantic>=2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
Pillow>=10.0
tqdm>=4.66.1
ftfy>=6.1.1
regex>=2023.10.3
timm>=0.9.12
